Move mock logic to separate file; automatically mock using decorators
